split out into `App.viewport.ts`
