+++ packages/excalidraw/components/EyeDropper.tsx
@@ -167,8 +167,8 @@ export const EyeDropper: React.FC<{
 
     // init color preview else it would show only after the first mouse move
     mouseMoveListener({
-      clientX: stableProps.app.lastViewportPosition.x,
-      clientY: stableProps.app.lastViewportPosition.y,
+      clientX: stableProps.app.viewport.lastPosition.x,
+      clientY: stableProps.app.viewport.lastPosition.y,
       altKey: false,
     });
